# Conflicts:
#	build.gradle
#	build/classes/java/main/me/aymanisam/hungergames/CommandDispatcher.class
#	build/classes/java/main/me/aymanisam/hungergames/HungerGames.class
#	build/classes/java/main/me/aymanisam/hungergames/commands/EndGameCommand.class
#	build/classes/java/main/me/aymanisam/hungergames/commands/LobbyReturnCommand.class
#	build/classes/java/main/me/aymanisam/hungergames/commands/SignSetCommand.class
#	build/classes/java/main/me/aymanisam/hungergames/handlers/GameSequenceHandler.class
#	build/classes/java/main/me/aymanisam/hungergames/handlers/ResetPlayerHandler.class
#	build/classes/java/main/me/aymanisam/hungergames/handlers/SetSpawnHandler.class
#	build/classes/java/main/me/aymanisam/hungergames/listeners/PlayerListener.class
#	build/classes/java/main/me/aymanisam/hungergames/listeners/SignClickListener.class
#	build/libs/Hungergames-1.7.4.jar
#	build/tmp/compileJava/previous-compilation-data.bin
#	src/main/java/me/aymanisam/hungergames/CommandDispatcher.java
#	src/main/java/me/aymanisam/hungergames/HungerGames.java
#	src/main/java/me/aymanisam/hungergames/commands/LobbyReturnCommand.java
#	src/main/java/me/aymanisam/hungergames/handlers/GameSequenceHandler.java

+++ src/main/java/me/aymanisam/hungergames/handlers/WorldResetHandler.java
@@ -9,27 +9,18 @@ import org.bukkit.block.EndGateway;
 import org.bukkit.block.ShulkerBox;
 import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
-import org.bukkit.entity.Player;
 import org.bukkit.persistence.PersistentDataContainer;
 import org.bukkit.persistence.PersistentDataType;
-import org.bukkit.scheduler.BukkitTask;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.logging.Level;
 
 public class WorldResetHandler {
     private final HungerGames plugin;
-    private final ArenaHandler arenaHandler;
-    private final ConfigHandler configHandler;
-    private final Map<String, BukkitTask> teleportTasks = new HashMap<>();
 
-    public WorldResetHandler(HungerGames plugin, LangHandler langHandler) {
+    public WorldResetHandler(HungerGames plugin) {
         this.plugin = plugin;
-        this.arenaHandler = new ArenaHandler(plugin, langHandler);
-        this.configHandler = plugin.getConfigHandler();
     }
 
     public void saveWorldState(World world) {
@@ -54,72 +45,32 @@ public class WorldResetHandler {
         });
     }
 
-    public void sendToWorld(World world) {
-        for (Player player : world.getPlayers()) {
-            String lobbyWorldName = (String) configHandler.createPluginSettings().get("lobby-world");
-            assert lobbyWorldName != null;
-            World lobbyWorld = Bukkit.getWorld(lobbyWorldName);
-            assert lobbyWorld != null;
-            player.teleport(lobbyWorld.getSpawnLocation());
-        }
-    }
-
     public void resetWorldState(World world) {
         File worldDirectory = world.getWorldFolder();
         File templateDirectory = new File(plugin.getDataFolder(), "templates" + File.separator + world.getName());
 
-        Bukkit.getScheduler().runTaskLater(plugin, () -> {
+        Bukkit.getScheduler().runTask(plugin, () -> {
             if (!templateDirectory.exists()) {
                 Bukkit.getLogger().severe("Template directory does not exist");
                 return;
             }
 
-            for (Entity entity : world.getEntities()) {
-                if (!(entity instanceof Player)) {
-                    entity.remove();
-                }
+            boolean unloaded = Bukkit.unloadWorld(world, false);
+            if (!unloaded) {
+                plugin.getLogger().log(Level.SEVERE, "Could not unload world");
+                return;
             }
 
-            arenaHandler.unloadChunks(world);
+            try {
+                FileUtils.deleteDirectory(worldDirectory);
+                FileUtils.copyDirectory(templateDirectory, worldDirectory);
+            } catch (IOException e) {
+                plugin.getLogger().log(Level.SEVERE, "Could not copy world folders");
+            }
 
-            Bukkit.getScheduler().runTaskLater(plugin, () -> {
-                try {
-                    boolean isUnloaded = plugin.getServer().unloadWorld(world, false);
-                    if (!isUnloaded) {
-                        Bukkit.getLogger().severe("World could not be unloaded");
-                    }
-
-                    FileUtils.deleteDirectory(worldDirectory);
-                    FileUtils.copyDirectory(templateDirectory, worldDirectory);
-
-                    loadWorld(world.getName());
-
-                    BukkitTask task = Bukkit.getScheduler().runTaskTimer(plugin, () -> {
-                        World reloadedWorld = Bukkit.getWorld(world.getName());
-                        if (reloadedWorld == null) {
-                            return;
-                        }
-
-                        for (Player player : world.getPlayers()) {
-                            player.teleport(reloadedWorld.getSpawnLocation());
-                            player.setGameMode(GameMode.ADVENTURE);
-                        }
-
-                        Bukkit.getScheduler().cancelTask(teleportTasks.get(world.getName()).getTaskId());
-                    }, 0L, 5L);
-
-                    teleportTasks.put(world.getName(), task);
-
-                } catch (IOException e) {
-                    Bukkit.getLogger().severe("An error occurred: " + e.getMessage());
-                }
-            }, 10L);
-        }, 10L);
-    }
-
-    public static void loadWorld(String worldName) {
-        WorldCreator creator = new WorldCreator(worldName);
-        Bukkit.createWorld(creator);
+            WorldCreator worldCreator = new WorldCreator(world.getName());
+            Bukkit.createWorld(worldCreator);
+        });
     }
 
     public void removeShulkers(World world) {
